requirements.txt
fastapi
uvicorn
yfinance
openai
tavily-python
python-dotenv
pydantic
pandas
